backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-appconf==1.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-rq==2.10.2
django-video-encoding==1.0.0
djangorestframework==3.15.2
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
redis==5.0.8
rq==1.16.2
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
times==0.7
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
zope.interface==7.1.0
